refactor(clones): split antiunify module
